Add unified batch processing API for OpenAI and Anthropic providers
- Introduced a new module for batch processing that provides a unified interface for handling batch requests across OpenAI and Anthropic.
- Implemented a type-safe Maybe/Result pattern for managing batch results with custom ID tracking.
- Added comprehensive data models, utility functions, and a BatchProcessor class to facilitate batch job management.
- Included example usage and documentation links for both providers to enhance usability and clarity.

This update significantly improves the batch processing capabilities and documentation, ensuring a streamlined experience for users working with multiple AI providers.
